backend/app/requirements.txt
uvicorn[standard]==0.30.6
fastapi==0.115.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
asyncpg==0.29.0
python-dotenv==1.0.1
